+++ changelog.md
@@ -4,6 +4,7 @@ Includes all important changes to the RPG adventure game.
 **v0.3 - Prerelease**
 - Added maps v0.1
 - Added Continuous Game Play
+- Added RPG classes (characters)
 
 **v0.2 - Prerelease**
 - Added pseudocode v0.1